--- src/DfE.ExternalApplications.Application/Templates/Commands/CreateCustomApplicationStatusCommand.cs
@@ -1,10 +1,0 @@
-using MediatR;
-using System;
-
-namespace DfE.ExternalApplications.Application.Templates.Commands
-{
-    public sealed record CreateCustomApplicationStatusCommand(
-        Guid TemplateId,
-        int ApplicationStatus,
-        string Label) : IRequest<Result<Guid>>;
-}
