--- src/DfE.ExternalApplications.Application/Templates/Commands/UpdateCustomApplicationStatusCommand.cs
@@ -1,10 +1,0 @@
-using MediatR;
-using System;
-
-namespace DfE.ExternalApplications.Application.Templates.Commands
-{
-    public sealed record UpdateCustomApplicationStatusCommand(
-        Guid CustomApplicationStatusId,
-        int ApplicationStatus,
-        string Label) : IRequest<Result>;
-}
